refactor: move config_schema to it's own package
 This is to avoid import conflicts and allow running
 the schema extraction without a full freqtrade installation.
