+++ build_helpers/extract_config_json_schema.py
@@ -9,7 +9,7 @@ import rapidjson
 def extract_config_json_schema():
     try:
         # Try to import from the installed package
-        from freqtrade.configuration.config_schema import CONF_SCHEMA
+        from freqtrade.config_schema import CONF_SCHEMA
     except ImportError:
         # If freqtrade is not installed, add the parent directory to sys.path
         # to import directly from the source
@@ -18,7 +18,7 @@ def extract_config_json_schema():
         sys.path.insert(0, str(freqtrade_dir))
 
         # Now try to import from the source
-        from freqtrade.configuration.config_schema import CONF_SCHEMA
+        from freqtrade.config_schema import CONF_SCHEMA
 
     schema_filename = Path(__file__).parent / "schema.json"
     with schema_filename.open("w") as f:
