+++ freqtrade/config_schema/__init__.py
@@ -1,0 +1,4 @@
+from freqtrade.config_schema.config_schema import CONF_SCHEMA
+
+
+__all__ = ["CONF_SCHEMA"]
