+++ freqtrade/config_schema/config_schema.py
